adiccion de alerta al guardar usuario

+++ js/index-form.js
@@ -10,53 +10,72 @@ if (idForm == null) {
   idForm = 0;
 }
 
-    cargarPagina();
+cargarPagina();
 
 function guardar() {
-  console.log("PRESIONO GUARDAR...");
-  let objUsuario = JSON.stringify({
-    idUsuario: idForm > 0 ? idForm : tablaUsuario.length + 1,
-    nombreApellido: document.getElementById("txtNombApellido").value,
-    Identificacion: document.getElementById("txtIdentificacion").value,
-    Telefono: document.getElementById("txtTelefono").value,
-    Direccion: document.getElementById("txtDireccion").value,
-    Estado: document.getElementById("cboEstado").value,
-  });
-  console.log(objUsuario);
-  //EDITAR
-  if (idForm > 0) {
-    for (const i in tablaUsuario){
-    let letUsuario = JSON.parse(tablaUsuario[i]);
-    if (letUsuario.idUsuario == idForm){
-        tablaUsuario[i] = objUsuario;
-        break;
-     }
+  Swal.fire({
+    title: "GUARDAR",
+    html: "DESEA GUARDAR LOS CAMBIOS?",
+    showDenyButton: true,
+    confirmButtonText: "SI",
+    denyButtonText: "NO",
+  }).then(
+    (result) => {
+    if (result.isConfirmed) {
+
+      console.log("PRESIONO GUARDAR...");
+      let objUsuario = JSON.stringify({
+        idUsuario: idForm > 0 ? idForm : tablaUsuario.length + 1,
+        nombreApellido: document.getElementById("txtNombApellido").value,
+        Identificacion: document.getElementById("txtIdentificacion").value,
+        Telefono: document.getElementById("txtTelefono").value,
+        Direccion: document.getElementById("txtDireccion").value,
+        Estado: document.getElementById("cboEstado").value,
+      });
+      console.log(objUsuario);
+      //EDITAR
+      if (idForm > 0) {
+        for (const i in tablaUsuario) {
+          let letUsuario = JSON.parse(tablaUsuario[i]);
+          if (letUsuario.idUsuario == idForm) {
+            tablaUsuario[i] = objUsuario;
+            break;
+          }
+        }
+      } else {
+        //NUEVOS PACIENTES
+        tablaUsuario.push(objUsuario);
+      }
+
+      localStorage.setItem("tablaUsuarioStorage", JSON.stringify(tablaUsuario));
+
+      Swal.fire('CAMBIOS  GUARDADOS','', 'Success').then(
+        (result) => {
+
+        window.location.replace("index.html");
+      }
+      
+      );
+    } else if(result.isdDenied) {
+      Swal.fire('CAMBIOS NO GUARDADOS','','info');
     }
-  } else {
-  //NUEVOS PACIENTES
-  tablaUsuario.push(objUsuario);
-
-  }
-  
-  localStorage.setItem("tablaUsuarioStorage", JSON.stringify(tablaUsuario));
-  window.location.replace("index.html");
-
- 
+  });
 }
 
 function cargarPagina() {
-    console.log("inicia funcion cargarPagina ");
+  console.log("inicia funcion cargarPagina ");
   if (idForm > 0) {
     for (const i in tablaUsuario) {
       let letUsuario = JSON.parse(tablaUsuario[i]);
-      if (letUsuario.idUsuario == idForm){
+      if (letUsuario.idUsuario == idForm) {
         document.getElementById("txtIdUsuario").value = letUsuario.idUsuario;
-        document.getElementById("txtNombApellido").value = letUsuario.nombreApellido;
-        document.getElementById("txtIdentificacion").value = letUsuario.Identificacion;
+        document.getElementById("txtNombApellido").value =
+          letUsuario.nombreApellido;
+        document.getElementById("txtIdentificacion").value =
+          letUsuario.Identificacion;
         document.getElementById("txtTelefono").value = letUsuario.Telefono;
         document.getElementById("txtDireccion").value = letUsuario.Direccion;
         break;
-        
       }
     }
   }
